requirements.txt
dearpygui==2.0.0
python-nmap==0.7.1
